feat: changes in the CRUD of memories to be able to execute each operation only by a registered user

+++ src/routes/memories.ts
@@ -3,8 +3,13 @@ import { z } from "zod"
 import { prisma } from "../lib/prisma"
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: { userId: request.user.sub },
       orderBy: {
         createdAt: "asc"
       }
@@ -18,7 +23,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -26,6 +31,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id }
     })
+
+    if (!memory.isPublic && memory.id !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
@@ -41,13 +51,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "d7db7ef1-5f0d-49fd-a4f0-b1b16958cf5f"
+        userId: request.user.sub
       }
     })
     return memory
   })
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -58,23 +68,40 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false)
     })
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-    const memory = await prisma.memory.update({
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
         coverUrl,
-        isPublic,
-        userId: "d7db7ef1-5f0d-49fd-a4f0-b1b16958cf5f"
+        isPublic
       }
     })
     return memory
   })
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
+
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id }
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
     await prisma.memory.delete({
       where: { id }
     })
